refactor to break up code into multiple files to make it easier to edit/maintain
