+++ src/main.rs
@@ -1,21 +1,24 @@
 // modules
 mod argc;
+mod can_i_connect;
 mod error;
-mod tests;
-mod types;
+mod helpers;
+mod integration_tests;
+mod options;
 
 // imports
 use self::error::Result;
-use crate::types::{CanIConnect, Options};
+use crate::can_i_connect::CanIConnect;
+use crate::options::Options;
 use argc::argc_app;
+use helpers::create_logger;
 use log::{error, info};
 use reqwest::Client;
 use std::time::Duration;
-use types::create_logger;
 
 #[tokio::main]
 async fn main() -> Result<()> {
-	// args setup
+	// options setup
 	let argc = argc_app().get_matches();
 	let options = match Options::from_argc(argc) {
 		Ok(options) => options,
@@ -26,24 +29,31 @@ async fn main() -> Result<()> {
 		.filter_level(options.log_level)
 		.init();
 
+	// can_i setup
+	let can_i_connect = CanIConnect {
+		http: options.http_hosts,
+		tcp: options.tcp_hosts,
+		timeout: options.timeout,
+		server_mode: !options.listen.is_empty(),
+		listen_addr: options.listen,
+		http_client: Some(
+			Client::builder()
+				.timeout(Duration::from_secs(options.timeout as u64))
+				.build()
+				.unwrap(),
+		),
+	};
+
 	// figure out if we are running in server mode (via --listen) or CLI mode
-	if options.listen != "" {
+	if can_i_connect.server_mode {
 		// we are in server mode
-		info!("Server mode Activated! Listening on: {}", options.listen);
+		info!(
+			"Server mode Activated! Listening on: {}",
+			can_i_connect.listen_addr
+		);
+		can_i_connect.bind().await;
 	} else {
 		// we are in CLI mode
-		// can_i setup
-		let can_i_connect = CanIConnect {
-			http: options.http_hosts,
-			tcp: options.tcp_hosts,
-			timeout: options.timeout,
-			http_client: Some(
-				Client::builder()
-					.timeout(Duration::from_secs(options.timeout as u64))
-					.build()
-					.unwrap(),
-			),
-		};
 		let connection_results = can_i_connect.connection_report().await;
 		info!(
 			"Successfully connected to [{}] hosts out of [{}] total hosts",
